pyproject.toml
[tool.poetry]
name = "inforce-test-task"
version = "0.1.0"
description = ""
authors = ["Dmytro Petliovanyi <petliovany@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pre-commit = "^3.4.0"
flake8 = "^6.1.0"
mypy = "^1.5.1"
isort = "^5.12.0"
django = "^4.2.5"
djangorestframework = "^3.14.0"
djangorestframework-jwt = "^1.11.0"
psycopg2-binary = "^2.9.7"
pytest = "^7.4.2"
pytest-django = "^4.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
